+++ tests/HelpScoutDocs/DocsApiClientTest.php
@@ -4,27 +4,28 @@ namespace HelpScoutDocs\Tests;
 
 use HelpScoutDocs\DocsApiClient;
 
-class DocsApiClientTest extends \PHPUnit_Framework_TestCase
+class DocsApiClientTest extends TestCase
 {
     /**
      * @var DocsApiClient
      */
     private $apiClient;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
         $this->apiClient = new DocsApiClient();
     }
-    
+
     /**
      * @test
-     * @expectedException \HelpScoutDocs\ApiException
-     * @expectedExceptionMessage Invalid API Key
      */
-    public function should_throw_an_exception_if_no_api_key_provided()
+    public function should_throw_an_exception_if_no_api_key_provided(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+        $this->expectExceptionMessage("Invalid API Key");
+
         $this->apiClient->getSites();
     }
 }