+++ tests/HelpScoutDocs/Endpoints/ArticlesTest.php
@@ -13,7 +13,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_return_articles_collection()
+    public function should_return_articles_collection(): void
     {
         $responseMock = $this->createResponseMock(200, __DIR__ . '/../../fixtures/articles/articles.json');
         $apiClient = $this->createTestApiClient($responseMock);
@@ -25,10 +25,11 @@ class ArticlesTest extends TestCase
 
     /**
      * @test
-     * @expectedException \HelpScoutDocs\ApiException
      */
-    public function should_throw_an_exception_on_invalid_file()
+    public function should_throw_an_exception_on_invalid_file(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+
         $responseMock = $this->createResponseMock(201, null);
         $apiClient = $this->createTestApiClient($responseMock);
 
@@ -40,7 +41,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_upload_an_article()
+    public function should_upload_an_article(): void
     {
         $responseMock = $this->createResponseMock(201, null);
         $apiClient = $this->createTestApiClient($responseMock);
@@ -57,7 +58,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_upload_an_article_and_respond_with_article_instance()
+    public function should_upload_an_article_and_respond_with_article_instance(): void
     {
         $responseMock = $this->createResponseMock(200, __DIR__ . '/../../fixtures/articles/uploaded.json');
         $apiClient = $this->createTestApiClient($responseMock);
@@ -75,7 +76,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_return_an_article_by_provided_id_or_number()
+    public function should_return_an_article_by_provided_id_or_number(): void
     {
         $responseMock = $this->createResponseMock(200, __DIR__ . '/../../fixtures/articles/article.json');
         $apiClient = $this->createTestApiClient($responseMock);
@@ -87,11 +88,12 @@ class ArticlesTest extends TestCase
 
     /**
      * @test
-     * @expectedException \HelpScoutDocs\ApiException
-     * @expectedExceptionCode 404
      */
-    public function should_throw_an_exception_if_invalid_article_id_or_number_provided()
+    public function should_throw_an_exception_if_invalid_article_id_or_number_provided(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+        $this->expectExceptionCode(404);
+
         $responseMock = $this->createResponseMock(404, null);
         $apiClient = $this->createTestApiClient($responseMock);
 
@@ -101,7 +103,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_search_for_articles_by_given_query_and_return_not_empty_result()
+    public function should_search_for_articles_by_given_query_and_return_not_empty_result(): void
     {
         $responseMock = $this->createResponseMock(200, __DIR__ . '/../../fixtures/articles/articles.json');
         $apiClient = $this->createTestApiClient($responseMock);
@@ -115,7 +117,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_return_related_articles_collection()
+    public function should_return_related_articles_collection(): void
     {
         $responseMock = $this->createResponseMock(200, __DIR__ . '/../../fixtures/articles/articles.json');
         $apiClient = $this->createTestApiClient($responseMock);
@@ -128,7 +130,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_return_article_revisions_collection()
+    public function should_return_article_revisions_collection(): void
     {
         $responseMock = $this->createResponseMock(200, __DIR__ . '/../../fixtures/articles/revisions.json');
         $apiClient = $this->createTestApiClient($responseMock);
@@ -141,7 +143,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_return_article_revision_by_given_id()
+    public function should_return_article_revision_by_given_id(): void
     {
         $responseMock = $this->createResponseMock(200, __DIR__ . '/../../fixtures/articles/revision.json');
         $apiClient = $this->createTestApiClient($responseMock);
@@ -153,11 +155,12 @@ class ArticlesTest extends TestCase
 
     /**
      * @test
-     * @expectedException \HelpScoutDocs\ApiException
-     * @expectedExceptionCode 404
      */
-    public function should_throw_an_exception_if_invalid_revision_id_provided()
+    public function should_throw_an_exception_if_invalid_revision_id_provided(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+        $this->expectExceptionCode(404);
+
         $responseMock = $this->createResponseMock(404, null);
         $apiClient = $this->createTestApiClient($responseMock);
 
@@ -167,7 +170,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_create_article_and_response_with_new_instance()
+    public function should_create_article_and_response_with_new_instance(): void
     {
         $responseMock = $this->createResponseMock(200, __DIR__ . '/../../fixtures/articles/article.json');
         $apiClient = $this->createTestApiClient($responseMock);
@@ -185,7 +188,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_create_article_and_assign_id()
+    public function should_create_article_and_assign_id(): void
     {
         $responseMock = $this->createResponseMock(200, null);
         $apiClient = $this->createTestApiClient($responseMock);
@@ -203,11 +206,12 @@ class ArticlesTest extends TestCase
 
     /**
      * @test
-     * @expectedException \HelpScoutDocs\ApiException
-     * @expectedExceptionCode 400
      */
-    public function should_throw_an_exception_if_malformed_article_provided()
+    public function should_throw_an_exception_if_malformed_article_provided(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+        $this->expectExceptionCode(400);
+
         $responseMock = $this->createResponseMock(400, null);
         $apiClient = $this->createTestApiClient($responseMock);
 
@@ -219,7 +223,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_update_existing_article_and_respond_with_updated_instance()
+    public function should_update_existing_article_and_respond_with_updated_instance(): void
     {
         $responseMock = $this->createResponseMock(200, __DIR__ . '/../../fixtures/articles/article.json');
         $apiClient = $this->createTestApiClient($responseMock);
@@ -238,7 +242,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_update_existing_article_and_respond_without_instance()
+    public function should_update_existing_article_and_respond_without_instance(): void
     {
         $responseMock = $this->createResponseMock(200, null);
         $apiClient = $this->createTestApiClient($responseMock);
@@ -257,18 +261,18 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_update_existing_article_view_count()
+    public function should_update_existing_article_view_count(): void
     {
         $responseMock = $this->createResponseMock(200, null);
         $apiClient = $this->createTestApiClient($responseMock);
-        
+
         $apiClient->updateViewCount(uniqid(), 1);
     }
 
     /**
      * @test
      */
-    public function should_delete_existing_article()
+    public function should_delete_existing_article(): void
     {
         $responseMock = $this->createResponseMock(200, null);
         $apiClient = $this->createTestApiClient($responseMock);
@@ -279,7 +283,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_save_existing_article_draft()
+    public function should_save_existing_article_draft(): void
     {
         $responseMock = $this->createResponseMock(200, null);
         $apiClient = $this->createTestApiClient($responseMock);
@@ -290,7 +294,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_delete_existing_article_draft()
+    public function should_delete_existing_article_draft(): void
     {
         $responseMock = $this->createResponseMock(200, null);
         $apiClient = $this->createTestApiClient($responseMock);
@@ -301,7 +305,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_return_articles_for_category()
+    public function should_return_articles_for_category(): void
     {
         $responseMock = $this->createResponseMock(200, __DIR__ . '/../../fixtures/articles/articles.json');
         $apiClient = $this->createTestApiClient($responseMock);
@@ -314,7 +318,7 @@ class ArticlesTest extends TestCase
     /**
      * @test
      */
-    public function should_return_articles_for_collection()
+    public function should_return_articles_for_collection(): void
     {
         $responseMock = $this->createResponseMock(200, __DIR__ . '/../../fixtures/articles/articles.json');
         $apiClient = $this->createTestApiClient($responseMock);