+++ tests/HelpScoutDocs/Endpoints/AssetsTest.php
@@ -11,28 +11,29 @@ class AssetsTest extends TestCase
     /**
      * @test
      */
-    public function should_create_article_asset()
+    public function should_create_article_asset(): void
     {
         $articleAsset = new ArticleAsset();
         $articleAsset->setArticleId(uniqid());
         $articleAsset->setAssetType(ArticleAsset::ARTICLE_ASSET_IMAGE);
         $articleAsset->setFile(__DIR__ . '/../../fixtures/assets/octocat.png');
-        
+
         $responseMock = $this->createResponseMock(201, __DIR__ . '/../../fixtures/assets/assets.json');
         $apiClient = $this->createTestApiClient($responseMock);
-        
+
         $created = $apiClient->createArticleAsset($articleAsset);
-        
+
         $this->assertInstanceOf(ArticleAsset::class, $created);
         $this->assertNotEmpty($created->getFileLink());
     }
 
     /**
      * @test
-     * @expectedException \HelpScoutDocs\ApiException
      */
-    public function should_throw_an_exception_about_article_asset_file()
+    public function should_throw_an_exception_about_article_asset_file(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+
         $articleAsset = new ArticleAsset();
 
         $responseMock = $this->createResponseMock(201, null);
@@ -43,10 +44,11 @@ class AssetsTest extends TestCase
 
     /**
      * @test
-     * @expectedException \HelpScoutDocs\ApiException
      */
-    public function should_throw_an_exception_about_missing_article_id()
+    public function should_throw_an_exception_about_missing_article_id(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+
         $articleAsset = new ArticleAsset();
         $articleAsset->setAssetType(ArticleAsset::ARTICLE_ASSET_IMAGE);
         $articleAsset->setFile(__DIR__ . '/../../fixtures/assets/octocat.png');
@@ -59,10 +61,11 @@ class AssetsTest extends TestCase
 
     /**
      * @test
-     * @expectedException \HelpScoutDocs\ApiException
      */
-    public function should_throw_an_exception_about_missing_article_asset_type()
+    public function should_throw_an_exception_about_missing_article_asset_type(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+
         $articleAsset = new ArticleAsset();
         $articleAsset->setArticleId(uniqid());
         $articleAsset->setFile(__DIR__ . '/../../fixtures/assets/octocat.png');
@@ -76,7 +79,7 @@ class AssetsTest extends TestCase
     /**
      * @test
      */
-    public function should_create_settings_asset()
+    public function should_create_settings_asset(): void
     {
         $settingsAsset = new SettingsAsset();
         $settingsAsset->setSiteId(uniqid());
@@ -94,10 +97,11 @@ class AssetsTest extends TestCase
 
     /**
      * @test
-     * @expectedException \HelpScoutDocs\ApiException
      */
-    public function should_throw_an_exception_about_settings_asset_file()
+    public function should_throw_an_exception_about_settings_asset_file(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+
         $settingsAsset = new SettingsAsset();
 
         $responseMock = $this->createResponseMock(201, null);
@@ -108,10 +112,11 @@ class AssetsTest extends TestCase
 
     /**
      * @test
-     * @expectedException \HelpScoutDocs\ApiException
      */
-    public function should_throw_an_exception_about_missing_settings_asset_type()
+    public function should_throw_an_exception_about_missing_settings_asset_type(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+
         $settingsAsset = new SettingsAsset();
         $settingsAsset->setSiteId(uniqid());
         $settingsAsset->setFile(__DIR__ . '/../../fixtures/assets/octocat.png');
@@ -124,10 +129,11 @@ class AssetsTest extends TestCase
 
     /**
      * @test
-     * @expectedException \HelpScoutDocs\ApiException
      */
-    public function should_throw_an_exception_about_missing_site_id()
+    public function should_throw_an_exception_about_missing_site_id(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+
         $settingsAsset = new SettingsAsset();
         $settingsAsset->setFile(__DIR__ . '/../../fixtures/assets/octocat.png');
         $settingsAsset->setAssetType(SettingsAsset::SETTINGS_ASSET_LOGO);
@@ -140,11 +146,12 @@ class AssetsTest extends TestCase
 
     /**
      * @test
-     * @expectedException \HelpScoutDocs\ApiException
-     * @expectedExceptionCode 400
      */
-    public function should_throw_an_exception_if_malformed_asset_provided()
+    public function should_throw_an_exception_if_malformed_asset_provided(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+        $this->expectExceptionCode(400);
+
         $articleAsset = new ArticleAsset();
         $articleAsset->setArticleId(uniqid());
         $articleAsset->setAssetType(ArticleAsset::ARTICLE_ASSET_IMAGE);
@@ -158,10 +165,11 @@ class AssetsTest extends TestCase
 
     /**
      * @tests
-     * @expectedException \HelpScoutDocs\ApiException
      */
-    public function should_throw_an_exception_if_api_key_is_invalid()
+    public function should_throw_an_exception_if_api_key_is_invalid(): void
     {
+        $this->expectException(\HelpScoutDocs\ApiException::class);
+
         $articleAsset = new ArticleAsset();
         $articleAsset->setArticleId(uniqid());
         $articleAsset->setAssetType(ArticleAsset::ARTICLE_ASSET_IMAGE);
@@ -170,7 +178,7 @@ class AssetsTest extends TestCase
         $responseMock = $this->createResponseMock(201, __DIR__ . '/../../fixtures/assets/assets.json');
         $apiClient = $this->createTestApiClient($responseMock);
         $apiClient->setApiKey('');
-        
+
         $apiClient->createArticleAsset($articleAsset);
     }
 }